moved to hospital quality filter menu to separate component
