+++ apps/skde/pages/sykehuskvalitet/index.tsx
@@ -1,6 +1,3 @@
-import { useEffect, useState } from "react";
-import { useRouter } from "next/router";
-import { StringParam, useQueryParam, withDefault } from "use-query-params";
 import AppBar from "@mui/material/AppBar";
 import Box from "@mui/material/Box";
 import CssBaseline from "@mui/material/CssBaseline";
@@ -9,65 +6,19 @@ import IconButton from "@mui/material/IconButton";
 import TuneIcon from "@mui/icons-material/Tune";
 import Toolbar from "@mui/material/Toolbar";
 import Typography from "@mui/material/Typography";
-import {
-  FilterMenu,
-  SelectedFiltersSection,
-  FilterSettingsValue,
-  FilterSettingsAction,
-} from "qmongjs";
 import Image from "next/image";
 import { imgLoader } from "qmongjs/src/helpers/functions";
+import { useState } from "react";
+import HospitalQualityFilterMenu from "./HospitalQualityFilterMenu";
 
 const drawerWidth = 300;
 
 /**
- * Extracts the query parameters from a path string
- *
- * @param path A path string
- * @returns A URLSearchParams object containing the query parameters in the path
- */
-export const extractQueryParams = (path: string) => {
-  const queryParamsStartIndex = path.lastIndexOf("?");
-  let queryString: string;
-
-  if (queryParamsStartIndex === -1) {
-    queryString = "";
-  } else {
-    queryString = path.substring(queryParamsStartIndex);
-  }
-
-  return new URLSearchParams(queryString);
-};
-
-/**
- * Sykehuskvalitet (hospital quality) page
+ * Hospital quality page (sykehuskvalitet)
  *
  * @returns The page component
  */
-export default function Sykehuskvalitet() {
-  // When the user navigates to the page, it may contain query parameters for
-  // filtering indicators. Use NextRouter to get the current path containing the
-  // initial query parameters.
-
-  const router = useRouter();
-  const path = router.asPath;
-
-  // Next's prerender stage causes problems for the initial values given to
-  // useReducer, because they are only set once by the reducer and are missing
-  // during Next's prerender stage. Tell FilterMenu to refresh its state during
-  // the first call after the prerender is done.
-
-  const [prevReady, setPrevReady] = useState(router.isReady);
-  const shouldRefreshInititalState = prevReady !== router.isReady;
-
-  useEffect(() => {
-    setPrevReady(router.isReady);
-  }, [router.isReady]);
-
-  const initialQueryParams = extractQueryParams(path);
-
-  // State and functions controlling the mobile drawer
-
+export default function HospitalQuality() {
   const [mobileOpen, setMobileOpen] = useState(false);
   const [isClosing, setIsClosing] = useState(false);
 
@@ -84,34 +35,6 @@ export default function Sykehuskvalitet() {
     if (!isClosing) {
       setMobileOpen(!mobileOpen);
     }
-  };
-
-  // Query parameters
-
-  const [selectedYear, setSelectedYear] = useQueryParam<string | undefined>(
-    "year",
-    withDefault(StringParam, undefined),
-  );
-
-  // Filter menu default values
-
-  const defaultYear = "2022";
-
-  /**
-   * Handler function for filter selection changes
-   */
-  const filterSelectionChanged = (
-    newFilterSettings: { map: Map<string, FilterSettingsValue[]> },
-    oldFilterSettings: { map: Map<string, FilterSettingsValue[]> },
-    action: FilterSettingsAction,
-  ) => {
-    setSelectedYear(newFilterSettings.map.get("selectedfilters")?.[0]?.value);
-    console.log(
-      "Filter selection changed",
-      newFilterSettings,
-      oldFilterSettings,
-      action,
-    );
   };
 
   return (
@@ -139,7 +62,7 @@ export default function Sykehuskvalitet() {
           </IconButton>
           <Box sx={{ marginLeft: 2 }}>
             <Typography variant="h6">
-              Resultater fra medisinske kvalitetsregistre
+              Sykehuskvalitet
             </Typography>
           </Box>
           <Box sx={{ marginLeft: "auto", marginTop: 1.15 }}>
@@ -174,23 +97,7 @@ export default function Sykehuskvalitet() {
             },
           }}
         >
-          <FilterMenu
-            refreshState={shouldRefreshInititalState}
-            onSelectionChanged={filterSelectionChanged}
-          >
-            <SelectedFiltersSection
-              accordion="false"
-              filterkey="selectedfilters"
-              sectionid="selectedfilters"
-              sectiontitle="Valgte filtre"
-              defaultvalues={[{ value: defaultYear, valueLabel: defaultYear }]}
-              initialselections={
-                selectedYear
-                  ? [{ value: selectedYear, valueLabel: selectedYear }]
-                  : undefined
-              }
-            />
-          </FilterMenu>
+          <HospitalQualityFilterMenu />
         </Drawer>
         <Drawer
           variant="permanent"
@@ -203,23 +110,7 @@ export default function Sykehuskvalitet() {
           }}
           open
         >
-          <FilterMenu
-            refreshState={shouldRefreshInititalState}
-            onSelectionChanged={filterSelectionChanged}
-          >
-            <SelectedFiltersSection
-              accordion="false"
-              filterkey="selectedfilters"
-              sectionid="selectedfilters"
-              sectiontitle="Valgte filtre"
-              defaultvalues={[{ value: defaultYear, valueLabel: defaultYear }]}
-              initialselections={
-                selectedYear
-                  ? [{ value: selectedYear, valueLabel: selectedYear }]
-                  : undefined
-              }
-            />
-          </FilterMenu>
+          <HospitalQualityFilterMenu />
         </Drawer>
       </Box>
       <Box
@@ -232,7 +123,7 @@ export default function Sykehuskvalitet() {
       >
         <Toolbar />
         <Typography paragraph>
-          Tabell for sykehuskvalitet vil vises her.
+          Resultater fra medisinske kvalitetsregistre
         </Typography>
       </Box>
     </Box>
