+++ packages/qmongjs/src/index.tsx
@@ -9,6 +9,7 @@ export { API_HOST } from "./components/RegisterPage";
 export { Layout } from "./components/Layout";
 export { level, levelSymbols } from "./helpers/functions";
 export { FilterMenu } from "./components/FilterMenu";
+export type { FilterMenuSelectionChangedHandler } from "./components/FilterMenu";
 export { FilterSettingsContext } from "./components/FilterMenu/FilterSettingsContext";
 export type { FilterSettingsValue } from "./components/FilterMenu/FilterSettingsContext";
 export {
